Clean up test structure
Move old test_utils content to test_helpers
to better correlate with current submodule naming
